requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Pydantic & Settings
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Database
sqlalchemy[asyncio]==2.0.38
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10

# Cache
redis[asyncio]==5.2.1

# Vector Store
faiss-cpu==1.9.0.post1
numpy==2.2.1

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.2.1

# HTTP Client
httpx==0.28.1
tenacity==9.0.0

# Background Scheduler
apscheduler==3.11.0

# Observability
prometheus-client==0.21.1
structlog==24.4.0

# LLM Clients
openai==1.59.9
anthropic==0.44.0

# Local Embeddings (optional, for dev without API keys)
sentence-transformers==3.3.1

# Utilities
python-dotenv==1.0.1
orjson==3.10.14
ujson==5.10.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
factory-boy==3.3.1
aiosqlite==0.20.0
